requirements.txt
numpy==1.26.0
openai==1.35.14
pysimplegui==5.0.6
soundcard==0.4.2
soundfile==0.12.1
loguru==0.7.2
pydub==0.25.1
simpleaudio==1.0.4
deepgram-sdk==3.*